Update color theme and links
This commit changes the color theme from a dark, black and purple color
theme to a light, white and blue color theme. The favicon has been
updated to reflect this.

This also removes the Snapchat link and adds the image, link, and
styling for Twitter.

Some old Snapchat files, such as the styling and logo image will stay in
case it is to be used later.